server/requirements.txt

Django==4.1.3
dj-database-url==0.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.2
whitenoise==5.3.0
coreapi==2.3.3
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.1
django-cors-headers==3.13.0

channels==3.0.4
channels-redis==3.2.0
daphne==3.0.2

Pillow==9.3.0
pytest-asyncio==0.20.1
pytest-django==4.5.2

pyautogui

apscheduler==3.9.1
django-dbbackup==4.0.2